:recycle: 인디케이터 컴포넌트 모듈 분리
